# Conflicts:
#	src/main/java/cn/nukkit/plugin/Plugin.java

+++ src/main/java/cn/nukkit/block/Stair.java
@@ -48,7 +48,7 @@ public abstract class Stair extends Transparent {
     @Override
     public boolean place(Item item, Block block, Block target, int face, double fx, double fy, double fz, Player player) {
         int[] faces = new int[]{0, 2, 1, 3};
-        this.meta = faces[player.getDirection()] & 0x03;
+        this.meta = (short) (faces[player.getDirection()] & 0x03);
         if ((fy > 0.5 && face != 1) || face == 0) {
             this.meta |= 0x04; //Upside-down stairs
         }
